+++ src/main/java/ro/ubbcluj/apm/am/AppointmentManagerApplication.java
@@ -1,18 +1,16 @@
 package ro.ubbcluj.apm.am;
 
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
+import ro.ubbcluj.apm.am.service.ContextService;
 
 public class AppointmentManagerApplication extends Application {
 
     @Override
-    public void start(Stage stage) throws Exception {
-        FXMLLoader fxmlLoader = new FXMLLoader(AppointmentManagerApplication.class.getResource("main-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), 320, 240);
+    public void start(Stage stage) {
         stage.setTitle("Appointment Manager");
-        stage.setScene(scene);
+        stage.setScene(new Scene(ContextService.getInstance().getMainView(), 800, 600));
         stage.show();
     }
 }